💄 :: 플레이어뷰 코드베이스로 변경
